Testing and tuning 3/7/24
Got a bug out that caused it to go haywire because of multiple motor top speeds and the class capping.

got 30s lap time. more work to do

Took out delays in PID (Line85)

Seems to have helped a ton in straight.

but we cannot corner. a lot of the issue seems to be frame width and lack of proper code for when the sensor loses the line.

Also, when it does loose the line. it goes haywire. we need to nail that down too.